added meme.js and message deletable checks

+++ commands/coins/getCoins.js
@@ -33,7 +33,7 @@ module.exports = {
                             addCoins(userMention)
 
                         function addCoins(usrID) {
-                            message.delete();
+                            if (message.deletable) message.delete();
                             coins.findOne({ guildID: guildID, userID: usrID }, (err, exists) => {
                                 if (!exists) return message.reply("User doesn't have coins yet.").then(m => m.delete(7500));
                                 if (exists.coins) return message.reply("User has: " + exists.coins + " coins").then(m => m.delete(7500));
